Splitting ArtistSearch ViewController
