+++ ArtistSearch/Modules/Artirst Search/UI/ViewController/ArtistSearchViewController+UICollectionViewDataSource.swift
@@ -1,0 +1,21 @@
+import UIKit
+
+extension ArtistSearchViewController: UICollectionViewDataSource {
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return searchData.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell: ArtistCollectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: ArtistCollectionViewCell.reusableIdentifier,
+                                                                                      for: indexPath) as? ArtistCollectionViewCell else {
+                                                                                        return UICollectionViewCell()
+        }
+        
+        let artist: Artist = searchData[indexPath.row]
+
+        cell.configureCell(with: artist)
+        
+        return cell
+    }
+}
