+++ ArtistSearch/Modules/Artirst Search/UI/ViewController/ArtistSearchViewController.swift
@@ -71,100 +71,3 @@ class ArtistSearchViewController: UIViewController, ArtistSearchViewProtocol {
         showInformationView(searchData.isEmpty , type: .noResults)
     }
 }
-
-// MARK: - UICollectionViewDelegate methods
-
-extension ArtistSearchViewController: UICollectionViewDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        if !searchData.isEmpty && searchData.indices.contains(indexPath.row) {
-            presenter?.presentArtistDetail(searchData[indexPath.row])
-        }
-    }
-}
-
-// MARK: - UICollectionViewDataSource methods
-
-extension ArtistSearchViewController: UICollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return searchData.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell: ArtistCollectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: ArtistCollectionViewCell.reusableIdentifier,
-                                                                                      for: indexPath) as? ArtistCollectionViewCell else {
-                                                                                        return UICollectionViewCell()
-        }
-        
-        let artist: Artist = searchData[indexPath.row]
-
-        cell.configureCell(with: artist)
-        
-        return cell
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout methods
-
-extension ArtistSearchViewController : UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        
-        let width: CGFloat = collectionView.bounds.width / 2 - 4
-        let height: CGFloat = 60.0
-        
-        return CGSize(width: width, height: height)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets.zero
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return 2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 10
-    }
-    
-}
-
-// MARK: - UISearchBarDelegate methods
-
-extension ArtistSearchViewController: UISearchBarDelegate {
-    
-    func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
-        searchBar.setShowsCancelButton(true, animated: true)
-        return true
-    }
-    
-    func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-        searchBar.text = ""
-        searchData = []
-        searchCollectionView.isHidden = true
-        searchBar.resignFirstResponder()
-        searchBar.setShowsCancelButton(false, animated: true)
-        showInformationView(true, type: .defaultInformation)
-    }
-    
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        guard let searchBarText = searchBar.text else {
-            searchData = []
-            searchCollectionView.isHidden = true
-            showInformationView(true, type: .defaultInformation)
-            return
-        }
-        searchBar.resignFirstResponder()
-        searchCollectionView.isHidden = false
-        presenter?.searchTerm(type: selectedFilterType, and: searchBarText)
-        showInformationView(false, type: .defaultInformation)
-    }
-}
